trouble with aligning comment stuff, gonna reformat this
