+++ msgThread/src/assets/exampleData/posts.json
@@ -1,0 +1,9 @@
+{
+    "GroupName": "Group name",
+    "TimeSincePost": "3",
+    "imageName": "Image.png",
+    "description": "Post description",
+    "likes": "21",
+    "commentNumber": "4",
+    "Name": "Helena"
+}
